refactor(codex): extract file-change accumulator from CodexNotificationRouter
The file_change normalization/merge helpers (pure) plus the per-item memo map
move to codexFileChangeAccumulator.ts (CodexFileChangeAccumulator owns the only
state — remember()/reset()); the shared asRecord/firstString coercers move to
codexNotificationHelpers.ts. The router keeps its emit methods so live emit
ordering + segment-reset timing are byte-for-byte unchanged; it just delegates
the merge/memo.

Behavior-preserving (full codex suite — 677 tests — passes untouched).
CodexNotificationRouter 879 -> 761 LOC; no new clones or cycles.
